src/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
python-jose
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
pydantic[email]
python-multipart
google-auth
requests
